♻️ (Move components) - Move components to the corresponding page in the app folder
