+++ src/app/image-generator/_components/Preview.tsx
