+++ src/app/image-generator/_components/SidebarSection.tsx
@@ -4,8 +4,8 @@ import {
   AccordionContent,
   AccordionItem,
   AccordionTrigger,
-} from "../ui/accordion";
-import { Button } from "../ui/button";
+} from "../../../components/ui/accordion";
+import { Button } from "../../../components/ui/button";
 
 interface ISidebarSectionProps {
   title: string;
