+++ src/app/image-generator/_components/section/background/RandomColor.tsx
