+++ src/app/image-generator/_components/section/background/TransparentButton.tsx
