+++ src/app/image-generator/_components/section/background/components/Gradient/GradientOrientationContainer.tsx
