+++ src/app/image-generator/_components/section/background/customColor/CustomColor.tsx
