+++ src/app/image-generator/_components/section/background/customColor/components/CustomGradientColor.tsx
@@ -2,8 +2,8 @@ import { Checkbox } from "@/components/ui/checkbox";
 import { Label } from "@/components/ui/label";
 import { defaultImageGeneratorSettings } from "@/lib/constant/defaultImageGeneratorSettings";
 import { useImageGeneratorStore } from "@/lib/store/imageGenerator.store";
-import GradientOrientationContainer from "../../components/gradient/GradientOrientationContainer";
-import RandomColor from "../../components/RandomColor";
+import GradientOrientationContainer from "../../components/Gradient/GradientOrientationContainer";
+import RandomColor from "../../RandomColor";
 import CustomColorPicker from "./CustomColorPicker";
 
 const CustomGradientColor = () => {
