+++ src/app/image-generator/_components/section/image/ImageBorder.tsx
