+++ src/app/image-generator/_components/section/image/ImageShadow.tsx
