+++ src/app/image-generator/_components/section/image/ImageVisibility.tsx
